remove toFhir method in testEvent for observation
