+++ backend/src/test/java/gov/cdc/usds/simplereport/api/converter/FhirConverterTest.java
@@ -16,16 +16,23 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.from;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
+import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.parser.IParser;
 import gov.cdc.usds.simplereport.db.model.PhoneNumber;
 import gov.cdc.usds.simplereport.db.model.Result;
 import gov.cdc.usds.simplereport.db.model.SupportedDisease;
 import gov.cdc.usds.simplereport.db.model.auxiliary.PhoneType;
 import gov.cdc.usds.simplereport.db.model.auxiliary.TestCorrectionStatus;
 import gov.cdc.usds.simplereport.db.model.auxiliary.TestResult;
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
+import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Stream;
+import org.apache.commons.io.IOUtils;
 import org.hl7.fhir.r4.model.ContactPoint;
 import org.hl7.fhir.r4.model.ContactPoint.ContactPointSystem;
 import org.hl7.fhir.r4.model.ContactPoint.ContactPointUse;
@@ -37,6 +44,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.skyscreamer.jsonassert.JSONAssert;
 import org.springframework.test.util.ReflectionTestUtils;
 
 class FhirConverterTest {
@@ -393,7 +401,7 @@ class FhirConverterTest {
 
   @Test
   void nullResult_convertToObservation() {
-    var actual = convertToObservation(null, TestCorrectionStatus.ORIGINAL, null);
+    var actual = convertToObservation((Result) null, TestCorrectionStatus.ORIGINAL, null);
 
     assertThat(actual).isNull();
   }
@@ -405,5 +413,76 @@ class FhirConverterTest {
             new Result(null, null, null, TestResult.POSITIVE), TestCorrectionStatus.ORIGINAL, null);
 
     assertThat(actual).isNull();
+  }
+
+  @Test
+  void multipleResults_toFhirObservation() throws IOException {
+    var covidId = "3c9c7370-e2e3-49ad-bb7a-f6005f41cf29";
+    var fluId = "302a7919-b699-4e0d-95ca-5fd2e3fcaf7a";
+    var covidResult =
+        new Result(null, new SupportedDisease("COVID-19", "96741-4"), TestResult.POSITIVE);
+    var fluResult =
+        new Result(null, new SupportedDisease("FLU A", "LP14239-5"), TestResult.NEGATIVE);
+    ReflectionTestUtils.setField(covidResult, "internalId", UUID.fromString(covidId));
+    ReflectionTestUtils.setField(fluResult, "internalId", UUID.fromString(fluId));
+
+    var actual =
+        convertToObservation(Set.of(covidResult, fluResult), TestCorrectionStatus.ORIGINAL, null);
+
+    FhirContext ctx = FhirContext.forR4();
+    IParser parser = ctx.newJsonParser();
+
+    assertThat(actual).hasSize(2);
+    String covidSerialized =
+        parser.encodeResourceToString(
+            actual.stream()
+                .filter(
+                    observation ->
+                        observation.getCode().getCodingFirstRep().is("http://loinc.org", "96741-4"))
+                .findFirst()
+                .orElse(null));
+    String fluSerialized =
+        parser.encodeResourceToString(
+            actual.stream()
+                .filter(
+                    observation ->
+                        observation
+                            .getCode()
+                            .getCodingFirstRep()
+                            .is("http://loinc.org", "LP14239-5"))
+                .findFirst()
+                .orElse(null));
+    var expectedSerialized1 =
+        IOUtils.toString(
+            Objects.requireNonNull(
+                getClass().getClassLoader().getResourceAsStream("fhir/observationCovid.json")),
+            StandardCharsets.UTF_8);
+    var expectedSerialized2 =
+        IOUtils.toString(
+            Objects.requireNonNull(
+                getClass().getClassLoader().getResourceAsStream("fhir/observationFlu.json")),
+            StandardCharsets.UTF_8);
+    JSONAssert.assertEquals(covidSerialized, expectedSerialized1, true);
+    JSONAssert.assertEquals(fluSerialized, expectedSerialized2, true);
+  }
+
+  @Test
+  void correction_toFhirObservation() throws IOException {
+    var id = "3c9c7370-e2e3-49ad-bb7a-f6005f41cf29";
+    var result = new Result(null, new SupportedDisease("COVID-19", "96741-4"), TestResult.NEGATIVE);
+    ReflectionTestUtils.setField(result, "internalId", UUID.fromString(id));
+
+    var actual = convertToObservation(Set.of(result), TestCorrectionStatus.CORRECTED, "woops");
+
+    FhirContext ctx = FhirContext.forR4();
+    IParser parser = ctx.newJsonParser();
+
+    String actualSerialized = parser.encodeResourceToString(actual.get(0));
+    var expectedSerialized1 =
+        IOUtils.toString(
+            Objects.requireNonNull(
+                getClass().getClassLoader().getResourceAsStream("fhir/observationCorrection.json")),
+            StandardCharsets.UTF_8);
+    JSONAssert.assertEquals(actualSerialized, expectedSerialized1, true);
   }
 }
